requirements.txt
fastapi>=0.95.0
python-telegram-bot==20.4
uvicorn
python-dotenv